requirements.txt
asgiref==3.7.2
attrs==23.1.0
autobahn==23.6.1
Automat==22.10.0
boto3==1.26.148
botocore==1.29.148
certifi==2023.5.7
cffi==1.15.1
channels==4.0.0
cloudinary==1.33.0
constantly==15.1.0
cryptography==41.0.1
daphne==4.0.0
dj-database-url==2.0.0
Django==4.2.1
django-cors-headers==4.0.0
django-rest-framework==0.1.0
django-storages==1.13.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
jmespath==1.0.1
Pillow==9.5.0
psycopg==3.1.9
psycopg2==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.7.0
pyOpenSSL==23.2.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
s3transfer==0.6.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.4
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.6.2
urllib3==1.26.16
whitenoise==6.4.0
zope.interface==6.0
